assets contain up to 64 bytes of name; growing memory arena starts with 1 MB allocated by default; textPack internationalized name is relative to folder; fixed monitor listing

+++ sources/cage-core/utility/threadSafeQueue.cpp
@@ -123,7 +123,7 @@ namespace cage
 		uint32 threadSafeQueuePriv::estimatedSize() const
 		{
 			threadSafeQueueImpl *impl = (threadSafeQueueImpl *)this;
-			return impl->items.size();
+			return numeric_cast<uint32>(impl->items.size());
 		}
 
 		holder<threadSafeQueuePriv> newThreadSafeQueue(const threadSafeQueueCreateConfig &config)
